Part-2/TaskScheduler/backend/pyproject.toml
[project]
name = "task-scheduler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ram Pata", email = "rpata@revspringinc.com" }
]
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.1.6",
    "email-validator>=2.0.0",
    "fastapi[standard]>=0.116.0",
    "greenlet>=3.2.4",
    "pydantic-settings>=2.7.1",
    "PyJWT>=2.8.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "alembic>=1.18.4",
    "asgi-lifespan>=2.1.0",
    "coverage>=7.13.5",
    "hypothesis>=6.140.0",
    "httpx>=0.28.1",
    "memory-profiler>=0.61.0",
    "mutmut>=3.4.0",
    "numpy>=2.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.1.0",
    "testcontainers[postgres]>=4.13.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
markers = [
    "container: tests that require docker/testcontainers",
    "mutation: tests used for mutation testing workflows",
]

[project.scripts]
task-scheduler = "app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
